Tarefa 5 - Criação das DTOs restantes

+++ Controllers/FilmeController.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using WebAPIDotNET5.DTOs;
 
 namespace WebAPIDotNET5.Controllers
 {
@@ -18,11 +19,24 @@ namespace WebAPIDotNET5.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult<Filme>> Post([FromBody] Filme filme)
+        public async Task<ActionResult<FilmeOutputPostDTO>> Post([FromBody] FilmeInputPostDTO filmeInputPostDto)
         {
+            if (filmeInputPostDto.Titulo == "") //TEMPORÁRIO
+            {
+                return NotFound("O Titulo do filme é obrigatório.");
+            }
+            else if (filmeInputPostDto.DiretorId == 0)
+            {
+                return NotFound("O Id do filme não pode ser 0.");
+            }
+
+            var diretorDoFilme = await _context.Diretores.FirstOrDefaultAsync(diretor => diretor.Id == filmeInputPostDto.DiretorId);
+            var filme = new Filme(filmeInputPostDto.Titulo, diretorDoFilme.Id);
             await _context.Filmes.AddAsync(filme);
             await _context.SaveChangesAsync();
-            return Ok(filme);
+            var filmeOutputPostDto = new FilmeOutputPostDTO(filme.Id, filme.Titulo);
+
+            return Ok(filmeOutputPostDto);
         }
 
         [HttpGet]
@@ -35,23 +49,48 @@ namespace WebAPIDotNET5.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<Filme>> Get(long id)
         {
+            if (id == 0)
+            {
+                return NotFound("O id do filme não pode ser 0.");
+            }
             var filme = await _context.Filmes.FirstOrDefaultAsync(filme => filme.Id == id);
             return Ok(filme);
         }
 
         [HttpPut("{id}")]
-        public async Task<ActionResult<Filme>> Put(long id, [FromBody] Filme filme)
+        public async Task<ActionResult<FilmeOutputPutDTO>> Put(long id, [FromBody] FilmeInputPutDTO filmeInputPutDto)
         {
+            if (id == 0)
+            {
+                return NotFound("O id do filme não pode ser 0.");
+            }
+            if (filmeInputPutDto.Titulo == "")
+            {
+                return NotFound("O titulo do filme é obrigatório.");
+            }
+
+            var filme = new Filme(filmeInputPutDto.Titulo, filmeInputPutDto.DiretorId);
             filme.Id = id;
             _context.Filmes.Update(filme);
             await _context.SaveChangesAsync();
-            return Ok(filme);
+            var filmeOutputPutDto = new FilmeOutputPutDTO(filme.Id, filme.Titulo);
+
+            return Ok(filmeOutputPutDto);
         }
 
         [HttpDelete("{id}")]
         public async Task<ActionResult<Filme>> Delete(long id)
         {
+            if (id == 0)
+            {
+                return NotFound("O id do filme não pode ser 0.");
+            }
             var filme = await _context.Filmes.FirstOrDefaultAsync(filme => filme.Id == id);
+
+            if (filme == null)
+            {
+                return NotFound("Filme não existe.");
+            }
             _context.Filmes.Remove(filme);
             await _context.SaveChangesAsync();
             return Ok(filme);
